requirements.txt
django
pillow
django-crispy-forms
djangorestframework
cx_oracle
social-auth-app-django
django-pwa